requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==1.7.1
pytz==2022.7
sqlparse==0.4.3
